+++ backend/src/domain-events/factory.js
@@ -6,8 +6,9 @@ const { DomainEventMetricsSubscriber } = require('./subscribers/metricsSubscribe
 const { DomainEventAuditSubscriber } = require('./subscribers/auditSubscriber');
 const { FutureWebSocketSubscriber } = require('./subscribers/futureWebSocketSubscriber');
 const { FutureAnalyticsSubscriber } = require('./subscribers/futureAnalyticsSubscriber');
+const { RedisPublisherSubscriber } = require('./subscribers/redisPublisherSubscriber');
 
-function createDomainEventBus({ repository = domainEventRepository, logger = null } = {}) {
+function createDomainEventBus({ repository = domainEventRepository, logger = null, redisPublisher = null } = {}) {
   const metricsSubscriber = new DomainEventMetricsSubscriber({ repository });
   const bus = new DomainEventBus({
     logger,
@@ -25,6 +26,9 @@ function createDomainEventBus({ repository = domainEventRepository, logger = nul
   bus.subscribe(new DomainEventPersistenceSubscriber({ repository }));
   bus.subscribe(metricsSubscriber);
   bus.subscribe(new DomainEventAuditSubscriber({ repository }));
+  if (redisPublisher) {
+    bus.subscribe(new RedisPublisherSubscriber({ publisher: redisPublisher }));
+  }
   bus.subscribe(new FutureWebSocketSubscriber());
   bus.subscribe(new FutureAnalyticsSubscriber());
   return bus;